requirements.txt
pdfplumber
python-docx
openai